+++ source/extensions/lightspeed.event.layers_cleanup/lightspeed/event/layers_cleanup/extension.py
@@ -9,7 +9,7 @@
 """
 import carb
 import omni.ext
-from lightspeed.events_manager.core import EVENTS_MANAGER_INSTANCE
+from lightspeed.events_manager import get_instance as _get_event_manager_instance
 from omni.flux.utils.common import reset_default_attrs as _reset_default_attrs
 
 from .core import EventLayersCleanupCore
@@ -28,9 +28,9 @@ class EventLayersCleanupExtension(omni.ext.IExt):
     def on_startup(self, ext_id):
         carb.log_info("[lightspeed.event.layers_cleanup] Lightspeed Event Layer Cleanup startup")
         self._core = EventLayersCleanupCore()
-        EVENTS_MANAGER_INSTANCE.register_event(self._core)
+        _get_event_manager_instance().register_event(self._core)
 
     def on_shutdown(self):
         carb.log_info("[lightspeed.event.layers_cleanup] Lightspeed Events Layer Cleanup shutdown")
-        EVENTS_MANAGER_INSTANCE.unregister_event(self._core)
+        _get_event_manager_instance().unregister_event(self._core)
         _reset_default_attrs(self)