+++ source/extensions/lightspeed.event.save_root_mod/lightspeed/event/save_root_mod/core.py
@@ -10,14 +10,14 @@
 import carb
 import omni.kit.usd.layers as _layers
 import omni.usd
-from lightspeed.events_manager.i_ds_event import ILSSEvent
+from lightspeed.events_manager import ILSSEvent as _ILSSEvent
 from lightspeed.layer_manager.core import LayerManagerCore, LayerType
 from omni.flux.utils.common import reset_default_attrs as _reset_default_attrs
 
 _CONTEXT = "/exts/lightspeed.event.save_root_mod/context"
 
 
-class EventSaveRootModCore(ILSSEvent):
+class EventSaveRootModCore(_ILSSEvent):
     def __init__(self):
         super().__init__()
         self.default_attr = {