+++ source/extensions/lightspeed.event.switch_to_replacement/lightspeed/event/switch_to_replacement/extension.py
@@ -9,7 +9,8 @@
 """
 import carb
 import omni.ext
-from lightspeed.events_manager.core import EVENTS_MANAGER_INSTANCE
+from lightspeed.events_manager import get_instance as _get_event_manager_instance
+from omni.flux.utils.common import reset_default_attrs as _reset_default_attrs
 
 from .core import SwitchToReplacementCore
 
@@ -27,20 +28,9 @@ class EventSaveRecentExtension(omni.ext.IExt):
     def on_startup(self, ext_id):
         carb.log_info("[lightspeed.event.switch_to_replacement] Lightspeed Event Switch To Replacement startup")
         self._core = SwitchToReplacementCore()
-        EVENTS_MANAGER_INSTANCE.register_event(self._core)
+        _get_event_manager_instance().register_event(self._core)
 
     def on_shutdown(self):
         carb.log_info("[lightspeed.event.switch_to_replacement] Lightspeed Events Switch To Replacement shutdown")
-        EVENTS_MANAGER_INSTANCE.unregister_event(self._core)
-        for attr, value in self.default_attr.items():
-            m_attr = getattr(self, attr)
-            if isinstance(m_attr, list):
-                m_attrs = m_attr
-            else:
-                m_attrs = [m_attr]
-            for m_attr in m_attrs:
-                destroy = getattr(m_attr, "destroy", None)
-                if callable(destroy):
-                    destroy()
-                del m_attr
-                setattr(self, attr, value)
+        _get_event_manager_instance().unregister_event(self._core)
+        _reset_default_attrs(self)