+++ source/extensions/lightspeed.events_manager/lightspeed/events_manager/core.py
@@ -8,33 +8,103 @@
 * license agreement from NVIDIA CORPORATION is strictly prohibited.
 """
 import typing
-from typing import List, Optional
+from typing import Any, Callable, List, Optional
 
+import carb
 from omni.flux.utils.common import Event as _Event
 from omni.flux.utils.common import EventSubscription as _EventSubscription
-from omni.flux.utils.common import reset_default_attrs as _reset_default_attrs
 
-if typing.TYPE_CHECKING:
-    from .i_ds_event import ILSSEvent
-
-EVENTS_MANAGER_INSTANCE = None
+if typing.TYPE_CHECKING:  # pragma: no cover
+    from .i_ds_event import ILSSEvent as _ILSSEvent
 
 
 class EventsManagerCore:
     """Manage events"""
 
-    def __init__(self, extension_path):
-        self.default_attr = {}
-        for attr, value in self.default_attr.items():
-            setattr(self, attr, value)
-
+    def __init__(self):
         self.__ds_events = []
+        self.__global_custom_events = {}
 
         self.__on_event_registered = _Event()
         self.__on_event_unregistered = _Event()
 
-        global EVENTS_MANAGER_INSTANCE
-        EVENTS_MANAGER_INSTANCE = self
+        self.__on_global_custom_event_registered = _Event()
+        self.__on_global_custom_event_unregistered = _Event()
+
+    def get_registered_global_event_names(self) -> List[str]:
+        """
+        Get a list of registered event(s) name(s)
+
+        Returns:
+            The list of event names
+        """
+        return list(self.__global_custom_events.keys())
+
+    def register_global_custom_event(self, name: str, show_warning: bool = False):
+        """
+        Register a global custom event
+
+        Args:
+            name: the name of your event you want to register
+            show_warning: show a warning or not if an event with the same name already exist
+        """
+        if name in self.__global_custom_events:
+            if show_warning:
+                carb.log_warn(f"Custom event {name} already exist")
+            return
+        self.__global_custom_events[name] = _Event()
+        self.__on_global_custom_event_registered(name)
+
+    def unregister_global_custom_event(self, name: str):
+        """
+        Unregister an event
+
+        Args:
+            name: the name of the event you want to unregister
+        """
+        if name in self.__global_custom_events:
+            del self.__global_custom_events[name]
+            self.__on_global_custom_event_unregistered(name)
+
+    def subscribe_global_custom_event(self, name: str, fn):
+        """
+        Return the object that will automatically unsubscribe when destroyed.
+        Called when we click on a tool (change of the selected tool)
+        """
+        if name not in self.__global_custom_events:
+            message = f"Custom event {name} doesn't exist"
+            carb.log_error(message)
+            raise ValueError(message)
+        return _EventSubscription(self.__global_custom_events[name], fn)
+
+    def subscribe_global_custom_event_register(self, fn: Callable[[str], Any]):
+        """
+        Return the object that will automatically unsubscribe when destroyed.
+        Called when we click on a tool (change of the selected tool)
+        """
+        return _EventSubscription(self.__on_global_custom_event_registered, fn)
+
+    def subscribe_global_custom_event_unregister(self, fn: Callable[[str], Any]):
+        """
+        Return the object that will automatically unsubscribe when destroyed.
+        Called when we click on a tool (change of the selected tool)
+        """
+        return _EventSubscription(self.__on_global_custom_event_unregistered, fn)
+
+    def call_global_custom_event(self, name: str, *args, **kwargs):
+        """
+        Call the registered event
+
+        Args:
+            name: the name of the event to call
+            *args: args that will be passed to the callbacks
+            **kwargs: kwargs that will be passed to the callbacks
+        """
+        if name not in self.__global_custom_events:
+            message = f"Custom event {name} doesn't exist"
+            carb.log_error(message)
+            raise ValueError(message)
+        self.__global_custom_events[name](*args, **kwargs)
 
     def _event_registered(self):
         """Call the event object that has the list of functions"""
@@ -47,7 +117,7 @@ class EventsManagerCore:
         """
         return _EventSubscription(self.__on_event_registered, fn)
 
-    def _event_unregistered(self, ds_event: "ILSSEvent"):
+    def _event_unregistered(self, ds_event: "_ILSSEvent"):
         """Call the event object that has the list of functions"""
         self.__on_event_unregistered(ds_event)
 
@@ -58,7 +128,7 @@ class EventsManagerCore:
         """
         return _EventSubscription(self.__on_event_unregistered, fn)
 
-    def register_event(self, ds_event: "ILSSEvent"):
+    def register_event(self, ds_event: "_ILSSEvent"):
         """
         Register a new event
         """
@@ -69,24 +139,25 @@ class EventsManagerCore:
     def get_registered_events(self) -> List:
         return self.__ds_events
 
-    def get_registered_event(self, name: str) -> Optional["ILSSEvent"]:
+    def get_registered_event(self, name: str) -> Optional["_ILSSEvent"]:
         for event in self.__ds_events:
             if event.name == name:
                 return event
         return None
 
-    def unregister_event(self, ds_event: "ILSSEvent"):
+    def unregister_event(self, ds_event: "_ILSSEvent"):
         """
-        Unregister a ILSSEvent
+        Unregister a _ILSSEvent
         """
-        ds_event.uninstall()
-        self._event_unregistered(ds_event)
-        self.__ds_events.remove(ds_event)
+        if ds_event in self.__ds_events:
+            ds_event.uninstall()
+            self._event_unregistered(ds_event)
+            self.__ds_events.remove(ds_event)
+        else:
+            carb.log_warn(f"Event {ds_event} was never registered")
 
     def destroy(self):
         for event in self.__ds_events:
             event.uninstall()
         self.__ds_events = []
-        _reset_default_attrs(self)
-        global EVENTS_MANAGER_INSTANCE
-        EVENTS_MANAGER_INSTANCE = None
+        self.__global_custom_events = {}