+++ source/extensions/lightspeed.events_manager/lightspeed/events_manager/tests/unit/__init__.py
@@ -1,5 +1,5 @@
 """
-* Copyright (c) 2021, NVIDIA CORPORATION.  All rights reserved.
+* Copyright (c) 2023, NVIDIA CORPORATION.  All rights reserved.
 *
 * NVIDIA CORPORATION and its licensors retain all intellectual property
 * and proprietary rights in and to this software, related documentation
@@ -7,8 +7,4 @@
 * distribution of this software and related documentation without an express
 * license agreement from NVIDIA CORPORATION is strictly prohibited.
 """
-__all__ = ["get_instance", "ILSSEvent"]
-
-
-from .extension import EventsManagerExtension, get_instance
-from .i_ds_event import ILSSEvent
+from .test_core import *