+++ source/extensions/lightspeed.trex.capture_tree.model/lightspeed/trex/capture_tree/model/model.py
@@ -8,23 +8,16 @@
 * license agreement from NVIDIA CORPORATION is strictly prohibited.
 """
 import asyncio
-import functools
-import re
-from typing import TYPE_CHECKING, Dict, List, Optional, Set, Tuple
+from typing import List, Optional, Tuple
 
-from lightspeed.common.constants import MATERIAL_RELATIONSHIP, REGEX_HASH, REGEX_MESH_PATH
 from lightspeed.trex.capture.core.shared import Setup as _CaptureCoreSetup
 from lightspeed.trex.replacement.core.shared import Setup as _ReplacementCoreSetup
 from omni import ui, usd
 from omni.flux.utils.common import Event as _Event
 from omni.flux.utils.common import EventSubscription as _EventSubscription
-from omni.flux.utils.common import async_wrap as _async_wrap
 from omni.flux.utils.common import reset_default_attrs as _reset_default_attrs
 
 from .items import CaptureTreeItem
-
-if TYPE_CHECKING:
-    from pxr import Sdf
 
 HEADER_DICT = {0: "Path", 1: "Progress"}
 
@@ -102,8 +95,7 @@ class CaptureTreeModel(ui.AbstractItemModel):
 
     def fetch_progress(self, items: Optional[List[CaptureTreeItem]] = None):
         self.cancel_tasks()
-        wrapped_fn = _async_wrap(functools.partial(self.__fetch_progress, items))
-        self._fetch_task = asyncio.ensure_future(wrapped_fn())
+        self._fetch_task = asyncio.ensure_future(self.__fetch_progress(items))
 
     def __on_stage_event(self, event):
         if event.type not in [int(usd.StageEventType.CLOSING), int(usd.StageEventType.CLOSED)]:
@@ -116,7 +108,16 @@ class CaptureTreeModel(ui.AbstractItemModel):
             self._fetch_task = None
         self._cancel_token = False
 
-    def __fetch_progress(self, items: Optional[List[CaptureTreeItem]] = None):
+    @usd.handle_exception
+    async def async_get_captured_hashes(self, item: CaptureTreeItem, replaced_items: List[str]):
+        replaced_result, all_assets_result = await self._core_capture.async_get_replaced_hashes(
+            item.path, replaced_items
+        )
+        item.replaced_items = len(replaced_result)
+        item.total_items = len(all_assets_result)
+
+    @usd.handle_exception
+    async def __fetch_progress(self, items: Optional[List[CaptureTreeItem]] = None):
         collection = items if items else self.__children
         # Reset the item state
         for item in collection:
@@ -134,46 +135,19 @@ class CaptureTreeModel(ui.AbstractItemModel):
             if self._cancel_token:
                 self.__task_completed()
                 return
-            replaced_items = replaced_items.union(self.__filter_hashes(layer))
+            replaced_items = replaced_items.union(_ReplacementCoreSetup.group_replaced_hashes(layer))
+
+        tasks = []
         for item in collection:
             # Allow cancelling the task for every capture item
             if self._cancel_token:
                 self.__task_completed()
                 return
-            # Update the replaced items values
-            captured_items = self.__filter_hashes(self._core_capture.get_captured_hashes(item.path))
-            item.replaced_items = len(captured_items & replaced_items)
-            item.total_items = len(captured_items)
+            tasks.append(asyncio.ensure_future(self.async_get_captured_hashes(item, replaced_items)))
+        await asyncio.wait(tasks, return_when=asyncio.ALL_COMPLETED)
         self._progress_updated()
         # Make sure the cancel is reset
         self.__task_completed()
-
-    def __filter_hashes(self, args: Tuple["Sdf.Layer", Dict[str, "Sdf.Path"]]) -> Set[str]:
-        """
-        Filter the hashes so that meshes and their associate materials count as a single entry
-        """
-        layer, hashes = args
-        filtered_hashes = set()
-        regex_mesh = re.compile(REGEX_MESH_PATH)
-        regex_hash = re.compile(REGEX_HASH)
-        for prim_hash, prim_path in hashes.items():
-            # Allow cancelling the task for every item
-            if self._cancel_token:
-                break
-            # If not a mesh, then no need to group
-            if not regex_mesh.match(str(prim_path)):
-                filtered_hashes.add(prim_hash)
-                continue
-            # If prim is a mesh, get the associated material instead to group them up
-            mesh_prim = layer.GetPrimAtPath(prim_path)
-            if MATERIAL_RELATIONSHIP in mesh_prim.relationships:
-                materials = mesh_prim.relationships[MATERIAL_RELATIONSHIP].targetPathList.explicitItems
-                # Always take the first material as there should never be more than 1 material here
-                match = regex_hash.match(str(materials[0]))
-                filtered_hashes.add(match.group(3) if match else prim_hash)
-            else:
-                filtered_hashes.add(prim_hash)
-        return filtered_hashes
 
     def _progress_updated(self):
         """Call the event object that has the list of functions"""