+++ source/extensions/lightspeed.common/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.5.0] - 2023-10-10
+### Add
+- Add global event key(s)
+
 ## [0.4.0] - 2023-09-11
 ### Add
 - Add `INGESTION_SCHEMA_PATHS`