+++ source/extensions/lightspeed.event.capture_persp_to_persp/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0] - 2023-10-10
+### Changed
+- cleanup
+- use custom global event to detect if the camera needs to be moved
+
 ## [1.0.1] - 2023-10-05
 ### Fixed
 - Fix flaky test