+++ source/extensions/lightspeed.events_manager/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.0] - 2023-10-10
+### Added
+- Add global custom event feature
+- Write tests
+
 ## [0.1.0] - 2021-11-12
-Created
+### Added
+- Created