+++ source/extensions/lightspeed.trex.capture.core.shared/docs/CHANGELOG.md
@@ -1,6 +1,15 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0] - 2023-10-10
+### Changed
+- import capture emit a custom global event now
+- optimize `get_hashes_from_capture_layer()`
+- added tests
+
+### Fixed
+- Fixed the "replaced" asset counter
+
 ## [1.0.0] - 2023-05-18
 ### Added
 - Created